Changed HCResultCollection to LinkedHashSet
Switching from ArrayList to LinkedHashSet allows me to access results (in the present case, cells) based on a reference (image name and ID) rather than needing to iterate through the list.  Using LikedHashSet rather than HashSet means the set retains the ordering of how it was added to the set.