Modified ParameterCollection
Hid the general workings of ParameterCollection in the method, rather than simply extending LinkedHashSet.  Now users only need to provide the module object as a reference instead of its hash.

Started using Parameter objects, which can store multiple fields.  Now have a "visible" parameter, which determines if the parameter window displays that parameter (as well as if it knows how to deal with that type of object).