Added GUI for module and parameter selection
GUI allows the user to add, remove and move modules.

Modules are shown in a list and can be selected, which shows their parameters.

Parameters are affected by the previous modules.

Still need to handle certain parameters and add image/folder input as well as running batch mode and testing modules.