Created custom OutputStream to display progress
Started work on custom OutputStream, which will write to a JTextField at the bottom of the GUI.  This is set for System.out, so is the default location all updates are written to (if verbose is selected).

Modified print statements in all modules, so they are now in the form "[module name] status".

Still need to create the text field to hold these (at the moment they're going to a dummy field).