Trimmed Maven dependencies
Moved Blob class to SegmentAnalysis, so I can remove a lot of extra dependencies.  This now just relies of a couple of Apache libraries (Math3 and IO) and the core ImageJ distribution.

The aim is to remove a lot of the dependency bloat that comes with adding Common to another project.