Updated CumStat to be single-valued
CumStat is no longer an array of values.  Instead, it calculates statistics for a single element.  Any extra dimensionality must be taken care of by the calling method.

Updated tests to reflect this.

Old version is still present as the deprecated "MultiCumStat" class, which will be removed soon.

Minor updates to NaNBinner to improve clarity of code.